tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/132.html

<!DOCTYPE html>
<html><head>
  <title>scheduler: external svg script </title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<div id="log"></div>
<script>
  var t = async_test();
  log("inline script #1")
</script>
<svg>
<script xlink:href="scripts/include-1.js"></script>
</svg>
<script>
log("inline script #2");
t.step(function() {
  assert_array_equals(eventOrder, ["inline script #1", "external script #1", "inline script #2"]);
  t.done();
});
</script>
